+++ loraham_kiss_tnc/build.sh
@@ -58,6 +58,7 @@ sources=(
   bridge.c
   bridge_conf.c
   bridge_kiss.c
+  bridge_rx.c
   bridge_tx_queue.c
   cli.c
   tcp_server.c
