+++ loraham_kiss_tnc/run_tests.sh
@@ -155,6 +155,7 @@ run_test bridge \
   "$script_dir/bridge.c" \
   "$script_dir/bridge_conf.c" \
   "$script_dir/bridge_kiss.c" \
+  "$script_dir/bridge_rx.c" \
   "$script_dir/bridge_tx_queue.c" \
   "$script_dir/ax25.c" \
   "$script_dir/tnc2.c" \
